Create CF_304A_PythagoreanTheoremII
I learned a better brute force approach to this problem. Instead of using three nested loops (one each for a,b,c and subsequently checking if a^2 + b^2 == c^2), this can be done in just two nested loops (one each for a and b). 

The key idea is to take sqrt(a^2 + b^2) and compare sqrt^2 with a^2 + b^2. If they are the same, then we have a Pythagorean triple! (a, b, sqrt(a^2 + b^2)). If not then do nothing.

I'll remember this technique from now on.