Create CF_472B_DesignTutorial
This is my first real B problem. I admit I had to see the solution. I was approaching it in such a bad way. I was just complicating matters when there existed a beautiful, elegant, short solution as above. Greedy algorithms are proving to be my Achilles's heel now. I learned to actually look at the sample test cases and understand their approach. 

But this is just beautiful.